add clear all functionality
