+++ Production/govuk_ios/Views/RecentlyViewed/RecentActivityView.swift
@@ -2,11 +2,26 @@ import SwiftUI
 import CoreData
 
 struct RecentActivityView: View {
-    @ObservedObject var viewModel: RecentActivityViewModel
+    @ObservedObject var viewModel: RecentActivitiesViewModel
     let lastVisitedFormatter = DateFormatter.recentActivityLastVisited
     @State private var showingAlert: Bool = false
+    let alertTitle = String.recentActivity.localized(
+        "recentActivityClearAllAlertTitle"
+    )
+    let alertDescription = String.recentActivity.localized(
+        "recentActivityClearAllAlertWarningDesc"
+    )
+    let alertPrimaryButtonTitle = String.recentActivity.localized(
+        "recentActivityAlertWarningConfirmation"
+    )
+    let alertSecondaryButtonTitle = String.recentActivity.localized(
+        "recentActivitiesAlertDismissText"
+    )
+    let toolbarButtonTitle = String.recentActivity.localized(
+        "recentActivityToolBarTitle"
+    )
 
-    init(viewModel: RecentActivityViewModel) {
+    init(viewModel: RecentActivitiesViewModel) {
         self.viewModel = viewModel
     }
 
@@ -14,7 +29,7 @@ struct RecentActivityView: View {
         ScrollView {
             if viewModel.model.todaysActivites.count >= 1 {
                 let rows = viewModel.model.todaysActivites.map({
-                    activityRow(activityItem: $0)
+                    viewModel.activityRow(activityItem: $0)
                 })
                 GroupedList(
                     content: [
@@ -30,7 +45,7 @@ struct RecentActivityView: View {
             }
             if viewModel.model.currentMonthActivities.count >= 1 {
                 let rows = viewModel.model.currentMonthActivities
-                    .map { activityRow(activityItem: $0) }
+                    .map { viewModel.activityRow(activityItem: $0) }
                 GroupedList(
                     content: [
                         GroupedListSection(
@@ -44,7 +59,7 @@ struct RecentActivityView: View {
                 )
             }
             if viewModel.model.recentMonthActivities.count >= 1 {
-                GroupedList(content: buildSectionsView())
+                GroupedList(content: viewModel.buildSectionsView())
             }
         }
         .toolbar {
@@ -52,51 +67,18 @@ struct RecentActivityView: View {
                 Button {
                     showingAlert.toggle()
                 } label: {
-                    Text("Clear")
+                    Text(toolbarButtonTitle)
                 }.alert(isPresented: $showingAlert, content: {
-                    Alert(title: Text("Are you sure you want to clear history"),
-                          dismissButton: Alert.Button.cancel(
-                            Text("Clear"), action: {
-                                withAnimation {
-                                    viewModel.deleteActivities()
-                                }
-                            }
-                          )
+                    Alert(
+                        title: Text(alertTitle),
+                        message: Text(alertDescription),
+                        primaryButton: .destructive(Text(alertPrimaryButtonTitle)) {
+                            viewModel.deleteActivities()
+                        },
+                        secondaryButton: .cancel()
                     )
                 })
             }
         }
-    }
-
-    private func activityRow(activityItem: ActivityItem) -> LinkRow {
-        LinkRow(
-            id: activityItem.id,
-            title: activityItem.title,
-            body: lastVisitedString(activity: activityItem),
-            action: {
-                viewModel.navigateToBrowser(item: activityItem)
-            }
-        )
-    }
-
-    private func buildSectionsView() -> [GroupedListSection] {
-        viewModel.model.recentMonthActivities.keys
-            .sorted { $0 > $1 }
-            .map {
-                let items = viewModel.model.recentMonthActivities[$0]
-                return GroupedListSection(
-                    heading: $0.title,
-                    rows: items?.map(activityRow) ?? [],
-                    footer: nil
-                )
-            }
-    }
-
-    private func lastVisitedString(activity: ActivityItem) -> String {
-        let copy = String.recentActivity.localized(
-            "recentActivityFormattedDateStringComponent"
-        )
-        let formattedDateString = lastVisitedFormatter.string(from: activity.date)
-        return "\(copy) \(formattedDateString)"
     }
 }
