+++ Tests/govuk_ios/govuk_ios_snapshot_tests/ReferenceImages_64/govuk_ios_snapshot_tests.AppUnavailableViewControllerSnapshotTests/test_loadInNavigationController_dark_rendersCorrectly@3x.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:34572383bfe5ddc4231d04472b6aefee91d2cea876a1e83b60b339d4bcb0ad13
-size 109108
+oid sha256:8fb965caecdf01a86d99b08eb78504b5dda97ec17905fcecfa29228e6cafb327
+size 108686
