+++ Tests/govuk_ios/govuk_ios_snapshot_tests/ReferenceImages_64/govuk_ios_snapshot_tests.AppUnavailableViewControllerSnapshotTests/test_loadInNavigationController_light_rendersCorrectly@3x.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f7b6bc08ac6944daa1a1e5222d382c56715d9bd104ee88c7e2b3f469f5b84557
-size 100995
+oid sha256:cf0a1bed60b649387410ffa04ac24a7227bbe188406c9f71aa13324a4d6909dc
+size 101349
