+++ Tests/govuk_ios/govuk_ios_snapshot_tests/ReferenceImages_64/govuk_ios_snapshot_tests.HomeViewControllerSnapshotTests/test_loadInNavigationController_dark_rendersCorrectly@3x.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:decbb4922abdc7a189163ccb88efc5e0184983b5e5ba87cdda88c9d717fd4a62
-size 155725
+oid sha256:757bd6287af47edaf9c89ff5be13c80b3da3ae9bafd06445a404aeb068f06520
+size 158340
