+++ Tests/govuk_ios/govuk_ios_snapshot_tests/ReferenceImages_64/govuk_ios_snapshot_tests.HomeViewControllerSnapshotTests/test_loadInNavigationController_light_rendersCorrectly@3x.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:99932602e83cf49ae48ef7759652b0c2264531eadb63e8a57cf3d2955031fb0b
-size 154025
+oid sha256:348117d4a1847b19d31a33f9e121304044c3bb91abd8f0cc83d37caa2e380514
+size 156563
