+++ Tests/govuk_ios/govuk_ios_snapshot_tests/ReferenceImages_64/govuk_ios_snapshot_tests.RecentActivityContainerViewSnapshotTests/test_loadInNavigationController_light_rendersCorrectly@3x.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:3f5dd4380c940a19c87f532a7ca2bee584d819b9dc1109445427be6fa2963de2
-size 103019
+oid sha256:75bd4772abdc965addf25aaaf801695b3a7c80e3d160659972909880d0884635
+size 109263
