+++ Tests/govuk_ios/govuk_ios_snapshot_tests/ReferenceImages_64/govuk_ios_snapshot_tests.SearchViewControllerSnapshotTests/test_search_failureResponse_genericError_rendersCorrectly@3x.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:83f340200204103d87adc3414293ab1fa15661f85df1e64bb98be7176f00f2bf
-size 115489
+oid sha256:f73d1c68b7de77dbbdf16fcaed62e05cdc933545f02a476eff5da65eb2fd7ffd
+size 117493
