+++ Tests/govuk_ios/govuk_ios_snapshot_tests/ReferenceImages_64/govuk_ios_snapshot_tests.SearchViewControllerSnapshotTests/test_search_failureResponse_networkUnavailable_rendersCorrectly@3x.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:5c51f210360b01a2cedf03a635295539265b31f8769c59545b3b58770a9804a1
-size 95601
+oid sha256:f3768ad0a06872ab61d152ce9fea43c3cc0d3a397118ae85686143b50a4d12db
+size 98825
